ACTIN-1666 Dockerfile for prediction
When built for a commit with a tag this should result in a Docker
container being pushed for the version of the tag.

Not clear to me what is useful other than having an environment with all
the required pips installed so doing that for now only. Also the tests
don't run properly yet so commenting them out.